// src/app/postlogin/dashboard/models/mkmyYojna.ts
export class MkmyYojnaPayLoad {
  consumerApplicationNo: any;
  address: any;
  area: any;
  consumerName: any;
  dcId: any;
  districtId: any;
  guardianName: any;
  khasra: any;
  khatoni: any;
  landAreaUnitId: any;
  loadRequested: any;
  loadRequestedId: any;
  natureOfWorkTypeId: any;
  pinCode: any;
  schemeTypeId: any;
  shortDescriptionOfWork: any;
  castCategory: any;
  mobilNo: any;
  samgraId:any;
  individualOrGroupId:any;
  samagraListDto:any
  aadharNo:any;
  




  // "natureOfWorkTypeId": 8,
  //   "address": "99, Barri Gujar ग्राम :बर्रीगुर्जर ग्राम पंचायत  :महोली जनपद पंचायत, बैरसिया जिला :भोपाल ",
  //   "area": 454,
  //   "castCategory": "GENERAL",
  //   "consumerName": "Ramnarayan Gurjar",
  //   "dcId": 170,
  //   "districtId": 6,
  //   "guardianName": "Parvat Singh",
  //   "khasra": "dftrterterer",
  //   "landAreaUnitId": 1,
  //   "loadRequested": "44",
  //   "loadRequestedId": 3,
  //   "pinCode": "462101",
  //   "schemeTypeId": 2,
  //   "shortDescriptionOfWork": "waoiefwelifwefiojeioj",
  //   "mobilNo": "8577915098",
  //   "samgraId": "106194976",
  //   "individualOrGroupId": 1,
  //   "samagraListDto": [],
  //   "aadharNo": "454543534545"









}